+++ tests/testthat/test-stats-htest.R
@@ -1,5 +1,8 @@
 context("stats-htest")
 
+skip_if_not_installed("modeltests")
+library(modeltests)
+
 test_that("htest tidier arguments", {
   check_arguments(tidy.htest)
   check_arguments(glance.htest)
@@ -28,15 +31,14 @@ test_that("tidy.htest/cor.test", {
   gl <- glance(pco)
   
   check_tidy_output(td)
-  check_glance_outputs(gl)
-  
+  check_glance_outputs(gl, strict = FALSE)
   
   sco <- suppressWarnings(cor.test(mtcars$mpg, mtcars$wt, method = "spearman"))
   td2 <- tidy(sco)
   gl2 <- glance(sco)
   
   check_tidy_output(td2)
-  check_glance_outputs(gl2)
+  check_glance_outputs(gl2, strict = FALSE)
 })
 
 test_that("tidy.htest/t.test", {
@@ -45,14 +47,13 @@ test_that("tidy.htest/t.test", {
   gl <- glance(tt)
   
   check_tidy_output(td)
-  check_glance_outputs(gl)
+  check_glance_outputs(gl, strict = FALSE)
 })
 
 test_that("tidy.htest/wilcox.test", {
   wt <- suppressWarnings(wilcox.test(mpg ~ am, mtcars))
   td <- tidy(wt)
   gl <- glance(wt)
-  
   
   check_tidy_output(td)
   check_glance_outputs(gl)
@@ -74,24 +75,26 @@ test_that("tidy.power.htest", {
   # gl <- glance(ptt)
   
   check_arguments(tidy.power.htest)
-  check_tidy_output(td)
+  check_tidy_output(td, strict = FALSE)
   # check_glance_outputs(gl). doesn't exist yet.
 })
 
 
 test_that("augment.htest (chi squared test)", {
-  check_arguments(augment.htest)
+  
+  # doesn't have a data argument
+  check_arguments(augment.htest, strict = FALSE)
   
   df <- as.data.frame(Titanic)
   tab <- xtabs(Freq ~ Sex + Class, data = df)
   
   chit <- chisq.test(tab) # 2D table
   au <- augment(chit)
-  check_tibble(au, method = "augment")
+  check_tibble(au, method = "augment", strict = FALSE)
   
   chit2 <- chisq.test(c(A = 20, B = 15, C = 25)) # 1D table
   au2 <- augment(chit2)
-  check_tibble(au2, method = "augment")
+  check_tibble(au2, method = "augment", strict = FALSE)
   
   tt <- t.test(rnorm(10))
   expect_error(